Remove yarn.lock
The yarn.lock was deleted because it was not necessary for the project and caused unexpected behavior on versioning. Yarn lockfile is used to lock down the versions of package dependencies to make sure all developers working on the project use the same versions. In this case, it is not necessary to keep the yarn.lock file.