Add Monaco Editor to the project
This commit adds the 'react-monaco-editor' package to our app
and creates a new MonacoEditorPage. This feature was added
to implement a comfortable and interactive coding environment within our application.
A new link and routing to the MonacoEditorPage was also added in home.tsx and App.tsx respectively.
The package-lock.json was automatically updated during the installation of the 'react-monaco-editor' package.